+++ rust/studio/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.1.2 - 2025-??-??
+
+- Built with rustc 1.91
+
 ## v0.1.1 - 2025-10-30
 
 - Built with rustc 1.90
@@ -9,3 +13,4 @@
 - Moved to manual rescanning to detect probes and USB devices.
 - Added network connectivity icon.
 - Single universal macOS dmg installer instead of separate Intel and Apple Silicon versions.
+- Added ability to load ROM config JSON files from disk in Create view.